+++ middleware-workflow-module/step/step-function/pom.xml
@@ -5,7 +5,7 @@
     <artifactId>step-function</artifactId>
     <version>1.0</version>
     <packaging>jar</packaging>
-    <name>Step lambda function</name>
+    <name>Workflow step lambda function</name>
     <properties>
         <maven.compiler.source>1.8</maven.compiler.source>
         <maven.compiler.target>1.8</maven.compiler.target>
