+++ middleware-workflow-module/step/step-function/src/test/java/co/tpg/workflow/function/StepFunctionTest.java
@@ -1,0 +1,21 @@
+package co.tpg.workflow.function;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertTrue;
+import org.junit.Test;
+
+public class StepFunctionTest {
+//  @Test
+//  public void successfulResponse() {
+//    App app = new App();
+//    GatewayResponse result = (GatewayResponse) app.handleRequest(null, null);
+//    assertEquals(result.getStatusCode(), 200);
+//    assertEquals(result.getHeaders().get("Content-Type"), "application/json");
+//    String content = result.getBody();
+//    assertNotNull(content);
+//    assertTrue(content.contains("\"message\""));
+//    assertTrue(content.contains("\"hello world\""));
+//    assertTrue(content.contains("\"location\""));
+//  }
+}
