+++ middleware-workflow-module/workflow-dao/src/main/java/co/tpg/workflow/function/dao/exception/BackendException.java
@@ -1,0 +1,16 @@
+package co.tpg.workflow.function.dao.exception;
+
+/**
+ * Exception class to decouple resources and backend.
+ * @author Rod
+ * @since 2019-10-06
+ */
+public class BackendException extends Exception {
+    public BackendException(String message) {
+        super(message);
+    }
+
+    public BackendException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
