--- middleware-workflow-module/workflow-model/src/main/java/co/tpg/workflow/exception/ProcessingException.java
@@ -1,11 +1,0 @@
-package co.tpg.workflow.exception;
-
-import lombok.*;
-
-@Builder
-@Getter
-@Setter
-@ToString
-public class ProcessingException extends Exception {
-    private String message;
-}
