+++ middleware-workflow-module/workflow-model/src/main/java/co/tpg/workflow/function/exception/FunctionWithException.java
@@ -1,0 +1,6 @@
+package co.tpg.workflow.function.exception;
+
+@FunctionalInterface
+public interface FunctionWithException<T, R, E extends Exception> {
+    R apply(T t) throws E;
+}
