+++ middleware-workflow-module/workflow-model/src/main/java/co/tpg/workflow/function/exception/ProcessingException.java
@@ -1,0 +1,11 @@
+package co.tpg.workflow.function.exception;
+
+import lombok.*;
+
+@Builder
+@Getter
+@Setter
+@ToString
+public class ProcessingException extends Exception {
+    private String message;
+}
