+++ middleware-workflow-module/workflow-model/src/main/java/co/tpg/workflow/function/request/HttpMethod.java
@@ -1,0 +1,17 @@
+package co.tpg.workflow.function.request;
+
+/**
+ * HTTP Constants class
+ * @author Rod
+ * @since 2019-09-04
+ */
+public final class HttpMethod {
+    public static final String GET = "GET";
+    public static final String HEAD = "HEAD";
+    public static final String POST = "POST";
+    public static final String PUT = "PUT";
+    public static final String PATCH = "PATCH";
+    public static final String DELETE = "DELETE";
+    public static final String OPTIONS = "OPTIONS";
+    public static final String TRACE = "TRACE";
+}
