+++ middleware-workflow-module/workflow-model/src/main/java/co/tpg/workflow/function/request/StepFieldRequest.java
@@ -1,0 +1,7 @@
+package co.tpg.workflow.function.request;
+
+import co.tpg.workflow.function.model.StepField;
+
+public class StepFieldRequest  extends AbstractRequest<StepField> {
+    public StepFieldRequest() {}
+}
