+++ middleware-workflow-module/workflow-model/src/main/java/co/tpg/workflow/function/request/StepRequest.java
@@ -1,0 +1,7 @@
+package co.tpg.workflow.function.request;
+
+import co.tpg.workflow.function.model.Step;
+
+public class StepRequest extends AbstractRequest<Step> {
+    public StepRequest() { }
+}
