+++ middleware-workflow-module/workflow-model/src/main/java/co/tpg/workflow/function/request/WorkflowRequest.java
@@ -1,0 +1,10 @@
+package co.tpg.workflow.function.request;
+
+import co.tpg.workflow.function.model.Workflow;
+
+/**
+ *
+ */
+public class WorkflowRequest extends AbstractRequest<Workflow> {
+    public WorkflowRequest() { }
+}
