+++ middleware-workflow-module/workflow-model/src/main/java/co/tpg/workflow/function/response/ErrorResponse.java
@@ -1,0 +1,11 @@
+package co.tpg.workflow.function.response;
+
+import co.tpg.workflow.function.exception.ProcessingException;
+
+/**
+ * Response class for General Errors.
+ * @author Rod
+ * @since 2019-09-28
+ */
+public class ErrorResponse extends AbstractResponse<ProcessingException> {
+}
