+++ middleware-workflow-module/workflow-model/src/main/java/co/tpg/workflow/function/response/StepFieldListResponse.java
@@ -1,0 +1,8 @@
+package co.tpg.workflow.function.response;
+
+import co.tpg.workflow.function.model.StepField;
+
+import java.util.List;
+
+public class StepFieldListResponse extends AbstractResponse<List<StepField>> {
+}
