+++ middleware-workflow-module/workflow-model/src/main/java/co/tpg/workflow/function/response/StepFieldResponse.java
@@ -1,0 +1,6 @@
+package co.tpg.workflow.function.response;
+
+import co.tpg.workflow.function.model.StepField;
+
+public class StepFieldResponse  extends AbstractResponse<StepField> {
+}
