+++ middleware-workflow-module/workflow-model/src/main/java/co/tpg/workflow/function/response/StepListResponse.java
@@ -1,0 +1,13 @@
+package co.tpg.workflow.function.response;
+
+import co.tpg.workflow.function.model.Step;
+
+import java.util.List;
+
+/**
+ * Response class for List of workflow step model.
+ * @author Andrej
+ * @since 2019-10-10
+ */
+public class StepListResponse extends AbstractResponse<List<Step>> {
+}
