+++ middleware-workflow-module/workflow-model/src/main/java/co/tpg/workflow/function/response/WorkflowResponse.java
@@ -1,0 +1,9 @@
+package co.tpg.workflow.function.response;
+
+import co.tpg.workflow.function.model.Workflow;
+
+/**
+ *
+ */
+public class WorkflowResponse extends AbstractResponse<Workflow> {
+}
