--- middleware-workflow-module/workflow-model/src/main/java/co/tpg/workflow/request/AbstractRequest.java
@@ -1,92 +1,0 @@
-package co.tpg.workflow.request;
-
-import java.util.Map;
-
-/**
- * Abstract class to provide request for lambda functions
- * @author Rod
- * @since 2019-09-04
- */
-public abstract class AbstractRequest<T> {
-    protected Map<String, String> queryStringParameters;
-    protected Map<String, String> pathParameters;
-    protected Map<String, String> headers;
-    protected String path;
-    protected String resourcePath;
-    protected String httpMethod;
-    protected String apiId;
-    protected String protocol;
-    protected T body;
-
-    public Map<String, String> getQueryStringParameters() {
-        return queryStringParameters;
-    }
-
-    public void setQueryStringParameters(Map<String, String> queryStringParameters) {
-        this.queryStringParameters = queryStringParameters;
-    }
-
-    public Map<String, String> getPathParameters() {
-        return pathParameters;
-    }
-
-    public void setPathParameters(Map<String, String> pathParameters) {
-        this.pathParameters = pathParameters;
-    }
-
-    public Map<String, String> getHeaders() {
-        return headers;
-    }
-
-    public void setHeaders(Map<String, String> headers) {
-        this.headers = headers;
-    }
-
-    public String getPath() {
-        return path;
-    }
-
-    public void setPath(String path) {
-        this.path = path;
-    }
-
-    public String getResourcePath() {
-        return resourcePath;
-    }
-
-    public void setResourcePath(String resourcePath) {
-        this.resourcePath = resourcePath;
-    }
-
-    public String getHttpMethod() {
-        return httpMethod;
-    }
-
-    public void setHttpMethod(String httpMethod) {
-        this.httpMethod = httpMethod;
-    }
-
-    public String getApiId() {
-        return apiId;
-    }
-
-    public void setApiId(String apiId) {
-        this.apiId = apiId;
-    }
-
-    public String getProtocol() {
-        return protocol;
-    }
-
-    public void setProtocol(String protocol) {
-        this.protocol = protocol;
-    }
-
-    public T getBody() {
-        return body;
-    }
-
-    public void setBody(T body) {
-        this.body = body;
-    }
-}
