--- middleware-workflow-module/workflow-model/src/main/java/co/tpg/workflow/request/StepFieldRequest.java
@@ -1,7 +1,0 @@
-package co.tpg.workflow.request;
-
-import co.tpg.workflow.model.StepField;
-
-public class StepFieldRequest  extends AbstractRequest<StepField> {
-    public StepFieldRequest() {}
-}
