--- middleware-workflow-module/workflow-model/src/main/java/co/tpg/workflow/request/StepRequest.java
@@ -1,7 +1,0 @@
-package co.tpg.workflow.request;
-
-import co.tpg.workflow.model.Step;
-
-public class StepRequest extends AbstractRequest<Step> {
-    public StepRequest() { }
-}
