--- middleware-workflow-module/workflow-model/src/main/java/co/tpg/workflow/response/ErrorResponse.java
@@ -1,11 +1,0 @@
-package co.tpg.workflow.response;
-
-import co.tpg.workflow.exception.ProcessingException;
-
-/**
- * Response class for General Errors.
- * @author Rod
- * @since 2019-09-28
- */
-public class ErrorResponse extends AbstractResponse<ProcessingException> {
-}
