--- middleware-workflow-module/workflow-model/src/main/java/co/tpg/workflow/response/StepFieldListResponse.java
@@ -1,8 +1,0 @@
-package co.tpg.workflow.response;
-
-import co.tpg.workflow.model.StepField;
-
-import java.util.List;
-
-public class StepFieldListResponse extends AbstractResponse<List<StepField>> {
-}
