--- middleware-workflow-module/workflow-model/src/main/java/co/tpg/workflow/response/StepFieldResponse.java
@@ -1,6 +1,0 @@
-package co.tpg.workflow.response;
-
-import co.tpg.workflow.model.StepField;
-
-public class StepFieldResponse  extends AbstractResponse<StepField> {
-}
