--- middleware-workflow-module/workflow-model/src/main/java/co/tpg/workflow/response/StepListResponse.java
@@ -1,12 +1,0 @@
-package co.tpg.workflow.response;
-
-import co.tpg.workflow.model.Step;
-import java.util.List;
-
-/**
- * Response class for List of workflow step model.
- * @author Andrej
- * @since 2019-10-10
- */
-public class StepListResponse extends AbstractResponse<List<Step>> {
-}
