--- middleware-workflow-module/workflow-model/src/main/java/co/tpg/workflow/response/StepResponse.java
@@ -1,6 +1,0 @@
-package co.tpg.workflow.response;
-
-import co.tpg.workflow.model.Step;
-
-public class StepResponse extends AbstractResponse<Step>{
-}
